SIGN-LINGO-AUTH/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
motor==3.3.1
pymongo==4.5.0
httpx==0.26.0
pydantic==2.6.1
pydantic-settings==2.1.0
python-dotenv==1.0.1
PyJWT==2.8.0
cloudinary==1.36.0
Werkzeug==2.3.7
email-validator==2.1.0
python-multipart==0.0.6
dnspython==2.6.1
certifi==2024.2.2